+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `parallel_runs` argument from `simulate_scenarios`, since parallelization
   can now be conveniently controlled via the new `Settings` mechanism
 - `empty_encoding` attribute from `SubspaceDiscrete`
+- `comp_rep` parameter from `SubspaceDiscrete.__init__`
 
 ### Deprecations
 - `set_random_seed` and `temporary_seed` utility functions